requirements.txt
streamlit
boto3
pandas
opencv-python-headless
Pillow